requirements.txt
gevent
gunicorn
fabric
markdown2
pymysql
django==2.1
djangorestframework
markdown 
django-filter
monthdelta
Pillow
xlrd
xlsxwriter